adding the ability to continue interviews
